// server/src/domain/entities/modelInterface/venue.interface.ts
import { Types } from "mongoose";

export interface IVenue {
    venueName?: string;
    rent?: number;
    capacity?: number;
    shift?: string;
    squareFeet?: number;
    timeSlots?: string[];
    availableDates?: string[];
    details?: string;
    features: string[];
    parkingFeatures: string[];
    venueDescription: string;
    terms: string;
    venueImages: string[];
    location: Types.ObjectId;
}